refactor: consolidate spin-specific build files
Using ublue-os/main and ublue-os/bluefin as references, builds for all
spins are now specified by a single Containerfile and a single workflow
file, configured via build arguments and `packages.json`. The Fedora
version for the generated images is taken out of the image name and
specified in the image's tags and metadata.